+++ DL_Track_US/__init__.py
@@ -5,7 +5,4 @@ __version__ = "0.1.2"
 __author__ = ["Paul Ritsche, Olivier Seynnes, Neil Cronin"]
 
 # original imports
-# from DL_Track_US.DL_Track_US_GUI import *
-
-# Carla imports
-from DL_Track_US_GUI import *
+from DL_Track_US.DL_Track_US_GUI import *