+++ DL_Track_US/gui_helpers/__init__.py
@@ -2,7 +2,6 @@ __all__ = [
     "calculate_architecture",
     "calibrate",
     "do_calculations",
-    "calibrate_video",
     "do_calculations_video",
     "calculate_architecture_video",
     "manual_tracing",
@@ -11,34 +10,34 @@ __all__ = [
     "data_augmentation",
     "curved_fascicles_functions",
     "do_calculations_curved",
+    "filter_data",
 ]
 
 __author__ = "Paul Ritsche, Olivier Seynnes, Neil Cronin"
 
 # original imports
-# from DL_Track_US.gui_helpers.calculate_architecture import *
-# from DL_Track_US.gui_helpers.calculate_architecture_video import *
-# from DL_Track_US.gui_helpers.calibrate import *
-# from DL_Track_US.gui_helpers.calibrate_video import *
-# from DL_Track_US.gui_helpers.curved_fascicles_functions import *
-# from DL_Track_US.gui_helpers.data_augmentation import *
-# from DL_Track_US.gui_helpers.do_calculations import *
-# from DL_Track_US.gui_helpers.do_calculations_video import *
-# from DL_Track_US.gui_helpers.file_analysis import *
-# from DL_Track_US.gui_helpers.manual_tracing import *
-# from DL_Track_US.gui_helpers.model_training import *
-# from DL_Track_US.gui_helpers.do_calculations_curved import *
+from DL_Track_US.gui_helpers.calculate_architecture import *
+from DL_Track_US.gui_helpers.calculate_architecture_video import *
+from DL_Track_US.gui_helpers.calibrate import *
+from DL_Track_US.gui_helpers.curved_fascicles_functions import *
+from DL_Track_US.gui_helpers.data_augmentation import *
+from DL_Track_US.gui_helpers.do_calculations import *
+from DL_Track_US.gui_helpers.do_calculations_video import *
+from DL_Track_US.gui_helpers.file_analysis import *
+from DL_Track_US.gui_helpers.manual_tracing import *
+from DL_Track_US.gui_helpers.model_training import *
+from DL_Track_US.gui_helpers.do_calculations_curved import *
+from DL_Track_US.gui_helpers.filter_data import *
 
 # Carla imports
-from gui_helpers.calculate_architecture import *
-from gui_helpers.calculate_architecture_video import *
-from gui_helpers.calibrate import *
-from gui_helpers.calibrate_video import *
-from gui_helpers.curved_fascicles_functions import *
-from gui_helpers.data_augmentation import *
-from gui_helpers.do_calculations import *
-from gui_helpers.do_calculations_curved import *
-from gui_helpers.do_calculations_video import *
-from gui_helpers.file_analysis import *
-from gui_helpers.manual_tracing import *
-from gui_helpers.model_training import *
+# from gui_helpers.calculate_architecture import *
+# from gui_helpers.calculate_architecture_video import *
+# from gui_helpers.calibrate import *
+# from gui_helpers.curved_fascicles_functions import *
+# from gui_helpers.data_augmentation import *
+# from gui_helpers.do_calculations import *
+# from gui_helpers.do_calculations_curved import *
+# from gui_helpers.do_calculations_video import *
+# from gui_helpers.file_analysis import *
+# from gui_helpers.manual_tracing import *
+# from gui_helpers.model_training import *